Documented source code. Renamed classes to more meaningful names.
