+++ Source/Scotec.Revit.Isolation/RevitCommandAvailabilityIsolationAttribute.cs
@@ -7,9 +7,14 @@ using System;
 namespace Scotec.Revit.Isolation;
 
 /// <summary>
-///     This attribute can be used to mark implementations
-///     of IExternalCommandAvailability" for execution in an isolated context.
+/// Represents an attribute that marks implementations of 
+/// <see cref="Autodesk.Revit.UI.IExternalCommandAvailability"/> for execution in an isolated context.
 /// </summary>
+/// <remarks>
+/// This attribute is intended to be applied to classes that implement 
+/// <see cref="Autodesk.Revit.UI.IExternalCommandAvailability"/> to ensure they are executed 
+/// in a specific isolated environment.
+/// </remarks>
 [AttributeUsage(AttributeTargets.Class)]
 public class RevitCommandAvailabilityIsolationAttribute : Attribute
 {
