+++ Source/Scotec.Revit/RevitDbApplicationIsolationAttribute.cs
@@ -8,9 +8,14 @@ using Autodesk.Revit.DB;
 namespace Scotec.Revit;
 
 /// <summary>
-///     This attribute can be used to mark implementations
-///     of <see cref="IExternalDBApplication"/> for execution in an isolated context.
+/// Represents an attribute used to designate implementations of 
+/// <see cref="IExternalDBApplication"/> for execution in an isolated context.
 /// </summary>
+/// <remarks>
+/// This attribute is marked as deprecated and will be removed in a future version. 
+/// It is recommended to reference the package <c>Scotec.Revit.Isolation</c> and use 
+/// the <c>Scotec.Revit.Isolation.RevitDbApplicationIsolation</c> attribute instead.
+/// </remarks>
 [AttributeUsage(AttributeTargets.Class)]
 [Obsolete("This attribute is marked as deprecated and will be removed in a future version. Reference package Scotec.Revit.Isolation and use the Scotec.Revit.Isolation.RevitDbApplicationIsolation attribute instead.")]
